modernize - refactor concurrent-all, replace tests
