+++ lib/concurrent-all.js
@@ -1,26 +1,40 @@
 'use strict';
 
-const invokeConcurrently = require('./invoke-concurrently.js');
-
-let ConcurrentAll;
-module.exports = ConcurrentAll = {
-    concurrentAll: function(functions, callback) {
-        invokeConcurrently(functions, ConcurrentAll._concurrentAllCallback, callback);
-    },
-    _concurrentAllCallback: function(currentIndex, allDone, error, result) {
-        if (error) {
-            this.numErrors = this.numErrors ? this.numErrors + 1 : 1;
+module.exports = {
+    concurrentAll: (functions, callback) => {
+        if (functions.length === 0) {
+            return callback(null, []);
         }
 
-        this.results[currentIndex] = {
-            error: error || null,
-            result: error ? null : result
+        const doConcurrently = (fn, index, state, done) => {
+            fn((error, result) => {
+                if (error) {
+                    state.numErrors = state.numErrors ? state.numErrors + 1 : 1;
+                }
+
+                state.results[index] = {
+                    error: error || null,
+                    result: error ? null : result
+                };
+
+                ++state.numComplete;
+
+                if (state.numComplete === state.numToDo) {
+                    return done(state.numErrors || null, state.results);
+                }
+            });
         };
 
-        this.numComplete++;
+        const initialState = {
+            numToDo: functions.length,
+            numErrors: null,
+            numComplete: 0,
+            results: [],
+            error: null
+        };
 
-        if (this.numComplete === this.numToDo) {
-            return process.nextTick(allDone.bind(null, this.numErrors === 0 ? null : this.numErrors, this.results));
-        }
+        functions.forEach((fn, index) => {
+            process.nextTick(() => doConcurrently(fn, index, initialState, callback));
+        });
     }
 };
