+++ lib/map-each.js
@@ -2,14 +2,13 @@
 
 module.exports = {
 
-    mapEach: (inputArray, forEachMethod, callback) => {
-        if (inputArray.length === 0) {
+    mapEach: (items, forEachItem, callback) => {
+        if (items.length === 0) {
             return callback(null, []);
         }
 
         const doMap = (fn, index, accumulator, terminateOnIndex, done) => {
-
-            fn(inputArray[index], (error, result) => {
+            fn(items[index], (error, result) => {
                 if (error) {
                     return done(error);
                 }
@@ -25,6 +24,6 @@ module.exports = {
             });
         };
 
-        doMap(forEachMethod, 0, [], inputArray.length, callback);
+        doMap(forEachItem, 0, [], items.length, callback);
     }
 };
